requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
ldap3==2.9.1
pyasn1==0.6.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
schedule==1.2.2
typing_extensions==4.12.2
urllib3==2.3.0
yandex-bot-py==1.0.6
